backend/requirements.txt
Django==5.0.7
django-ninja==1.2.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2-binary~=2.9.9
uvicorn==0.30.1